feat(6648): extract governance primitives into kamn-governance
