+++ crates/kamn-core/Cargo.toml
@@ -16,6 +16,7 @@ libp2p-live-transport = ["dep:libp2p", "dep:tokio"]
 kamn-bridges = { path = "../kamn-bridges" }
 kamn-crypto = { path = "../kamn-crypto" }
 kamn-data-layer = { path = "../kamn-data-layer" }
+kamn-governance = { path = "../kamn-governance" }
 kamn-kolme = { path = "../kamn-kolme" }
 kamn-live-probe-matrix = { path = "../kamn-live-probe-matrix" }
 kamn-runtime-guards = { path = "../kamn-runtime-guards" }
