+++ crates/kamn-governance/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "kamn-governance"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[lints]
+workspace = true
